moving components to form-layout directory
