+++ docs/reference/va-forms-system-core.formtitleprops.md
@@ -1,0 +1,25 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@department-of-veterans-affairs/va-forms-system-core](./va-forms-system-core.md) &gt; [FormTitleProps](./va-forms-system-core.formtitleprops.md)
+
+## FormTitleProps interface
+
+FormTitle properties
+
+<b>Signature:</b>
+
+```typescript
+export interface FormTitleProps 
+```
+
+## Example
+
+\`\`\` { title: 'My Example Title' subTitle: 'My Example Sub Title for more information' }
+
+## Properties
+
+|  Property | Type | Description |
+|  --- | --- | --- |
+|  [subTitle?](./va-forms-system-core.formtitleprops.subtitle.md) | string | <i>(Optional)</i> |
+|  [title](./va-forms-system-core.formtitleprops.title.md) | string |  |
+
