+++ docs/reference/va-forms-system-core.formtitleprops.subtitle.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@department-of-veterans-affairs/va-forms-system-core](./va-forms-system-core.md) &gt; [FormTitleProps](./va-forms-system-core.formtitleprops.md) &gt; [subTitle](./va-forms-system-core.formtitleprops.subtitle.md)
+
+## FormTitleProps.subTitle property
+
+<b>Signature:</b>
+
+```typescript
+subTitle?: string;
+```
