+++ docs/reference/va-forms-system-core.formtitleprops.title.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@department-of-veterans-affairs/va-forms-system-core](./va-forms-system-core.md) &gt; [FormTitleProps](./va-forms-system-core.formtitleprops.md) &gt; [title](./va-forms-system-core.formtitleprops.title.md)
+
+## FormTitleProps.title property
+
+<b>Signature:</b>
+
+```typescript
+title: string;
+```
