+++ docs/reference/va-forms-system-core.introductionpage.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@department-of-veterans-affairs/va-forms-system-core](./va-forms-system-core.md) &gt; [IntroductionPage](./va-forms-system-core.introductionpage.md)
+
+## IntroductionPage variable
+
+<b>Signature:</b>
+
+```typescript
+IntroductionPage: () => JSX.Element
+```
