+++ src/form-builder/index.ts
@@ -4,7 +4,7 @@ export { default as CheckboxFieldGroup } from './CheckboxFieldGroup';
 export { default as SelectField } from './SelectField';
 export { default as DateField } from './DateField';
 export { default as EmailField } from './EmailField';
-export { default as FormTitle } from './FormTitle';
+export { default as FormTitle } from '../form-layout/FormTitle';
 export { default as OMBInfo } from './OMBInfo';
 export { RadioGroup } from './RadioGroup';
 export { default as FullNameField } from './FullNameField';
