+++ src/form-builder/types.ts
@@ -37,21 +37,6 @@ export type CheckboxGroupProps = FieldProps<string> & {
 };
 
 /**
- * FormTitle properties
- *
- * @example
- * ```
- * {
- *    title: 'My Example Title'
- *    subTitle: 'My Example Sub Title for more information'
- * }
- */
-export interface FormTitleProps {
-  title: string;
-  subTitle?: string;
-}
-
-/**
  * OMBInfo properties
  *
  * @example
