+++ src/form-layout/index.ts
@@ -1,0 +1,3 @@
+export { default as FormTitle } from './FormTitle';
+export { default as IntroductionPage } from './IntroductionPage';
+export * from './types';
