+++ src/form-layout/types.ts
@@ -1,0 +1,14 @@
+/**
+ * FormTitle properties
+ *
+ * @example
+ * ```
+ * {
+ *    title: 'My Example Title'
+ *    subTitle: 'My Example Sub Title for more information'
+ * }
+ */
+export interface FormTitleProps {
+  title: string;
+  subTitle?: string;
+}
