+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './form-builder';
+export * from './form-layout';
 export * from './routing';
 
 export { default as DebuggerView } from './debugger-view';
